refactor mock version
