+++ voicevox_engine/tts_pipeline/song_engine.py
@@ -5,12 +5,11 @@ from typing import Any, Final, Literal, TypeAlias
 import numpy as np
 from numpy.typing import NDArray
 
-from voicevox_engine.utility.core_version_utility import get_latest_version
-
 from ..core.core_adapter import CoreAdapter, DeviceSupport
-from ..core.core_initializer import MOCK_VER, CoreManager
+from ..core.core_initializer import CoreManager
 from ..core.core_wrapper import CoreWrapper
 from ..metas.Metas import StyleId
+from ..utility.core_version_utility import MOCK_VER, get_latest_version
 from .audio_postprocessing import raw_wave_to_output_wave
 from .model import (
     FrameAudioQuery,
@@ -442,8 +441,6 @@ class SongEngineManager:
 
 def make_song_engines_from_cores(core_manager: CoreManager) -> SongEngineManager:
     """コア一覧からSongエンジン一覧を生成する"""
-    # FIXME: `MOCK_VER` を循環 import 無しに `initialize_cores()` 関連モジュールから import する
-    MOCK_VER = "0.0.0"
     song_engines = SongEngineManager()
     for ver, core in core_manager.items():
         if ver == MOCK_VER:
