+++ voicevox_engine/utility/core_version_utility.py
@@ -1,6 +1,10 @@
 """バージョンに関する utility"""
 
+from typing import Final
+
 from semver.version import Version
+
+MOCK_VER: Final = "0.0.0"
 
 
 def get_latest_version(versions: list[str]) -> str:
